# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type SignupBodyRole =
  (typeof SignupBodyRole)[keyof typeof SignupBodyRole];

export const SignupBodyRole = {
  customer: "customer",
  vendor: "vendor",
} as const;

export interface SignupBody {
  email: string;
  password: string;
  name: string;
  role: SignupBodyRole;
  cartName?: string;
  category?: string;
  city?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  customer: "customer",
  vendor: "vendor",
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  role: UserProfileRole;
  createdAt: string;
}

export type VendorSummarySubscriptionStatus =
  (typeof VendorSummarySubscriptionStatus)[keyof typeof VendorSummarySubscriptionStatus];

export const VendorSummarySubscriptionStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface VendorSummary {
  id: number;
  cartName: string;
  category: string;
  city: string;
  coverPhoto?: string;
  startingPrice?: number;
  avgRating?: number;
  isActive: boolean;
  subscriptionStatus: VendorSummarySubscriptionStatus;
  onboardingComplete: boolean;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
  vendorProfile?: VendorSummary;
}

export interface MeResponse {
  user: UserProfile;
  vendorProfile?: VendorSummary;
}

export type VendorDetailSubscriptionStatus =
  (typeof VendorDetailSubscriptionStatus)[keyof typeof VendorDetailSubscriptionStatus];

export const VendorDetailSubscriptionStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface PricingPackage {
  name: string;
  price: number;
  description: string;
}

export interface VendorDetail {
  id: number;
  userId: number;
  cartName: string;
  category: string;
  bio?: string;
  city: string;
  coverPhoto?: string;
  galleryPhotos?: string[];
  startingPrice?: number;
  packages?: PricingPackage[];
  isActive: boolean;
  subscriptionStatus: VendorDetailSubscriptionStatus;
  avgRating: number;
  profileViews: number;
  totalReviews: number;
  onboardingComplete: boolean;
  isSaved?: boolean;
}

export interface VendorListResponse {
  vendors: VendorSummary[];
  total: number;
}

export interface VendorDetailResponse {
  vendor: VendorDetail;
}

export interface UpdateVendorBody {
  cartName?: string;
  category?: string;
  bio?: string;
  city?: string;
  coverPhoto?: string;
  galleryPhotos?: string[];
  startingPrice?: number;
  packages?: PricingPackage[];
  isActive?: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  declined: "declined",
} as const;

export interface Booking {
  id: number;
  customerId: number;
  vendorId: number;
  eventDate: string;
  eventType: string;
  guestCount: number;
  location: string;
  message?: string;
  status: BookingStatus;
  vendorNote?: string;
  createdAt: string;
  customerName?: string;
  vendorCartName?: string;
  hasReview?: boolean;
}

export interface CreateBookingBody {
  vendorId: number;
  eventDate: string;
  eventType: string;
  guestCount: number;
  location: string;
  message?: string;
}

export type RespondBookingBodyStatus =
  (typeof RespondBookingBodyStatus)[keyof typeof RespondBookingBodyStatus];

export const RespondBookingBodyStatus = {
  confirmed: "confirmed",
  declined: "declined",
} as const;

export interface RespondBookingBody {
  status: RespondBookingBodyStatus;
  vendorNote?: string;
}

export interface BookingResponse {
  booking: Booking;
}

export interface BookingListResponse {
  bookings: Booking[];
}

export interface Review {
  id: number;
  bookingId: number;
  customerId: number;
  vendorId: number;
  rating: number;
  body?: string;
  createdAt: string;
  customerName?: string;
}

export interface CreateReviewBody {
  bookingId: number;
  rating: number;
  body?: string;
}

export interface ReviewResponse {
  review: Review;
}

export interface ReviewListResponse {
  reviews: Review[];
}

export interface SavedVendorsResponse {
  vendors: VendorSummary[];
}

export interface ToggleSavedResponse {
  saved: boolean;
}

export type SubscriptionResponseSubscriptionStatus =
  (typeof SubscriptionResponseSubscriptionStatus)[keyof typeof SubscriptionResponseSubscriptionStatus];

export const SubscriptionResponseSubscriptionStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface SubscriptionResponse {
  success: boolean;
  subscriptionStatus: SubscriptionResponseSubscriptionStatus;
}

export interface VendorStatsResponse {
  totalBookings: number;
  pendingRequests: number;
  profileViews: number;
  avgRating: number;
  confirmedBookings: number;
  declinedBookings: number;
}

export interface OnboardingBody {
  cartName: string;
  category: string;
  bio: string;
  city: string;
  coverPhoto?: string;
  galleryPhotos?: string[];
  startingPrice: number;
  packages?: PricingPackage[];
  activateSubscription?: boolean;
}

export type ListVendorsParams = {
  category?: string;
  city?: string;
  minPrice?: number;
  maxPrice?: number;
  rating?: number;
  availableOn?: string;
  limit?: number;
  offset?: number;
};