# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cartly vendor marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
...
  /bookings/{id}/respond:
    put:
      operationId: respondToBooking
      tags: [bookings]
      summary: Vendor accepts or declines a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondBookingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review after a confirmed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
  /reviews/vendor/{vendorId}:
    get:
      operationId: getVendorReviews
      tags: [reviews]
      summary: Get reviews for a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
  /saved/{vendorId}:
    post:
      operationId: toggleSavedVendor
      tags: [saved]
      summary: Toggle save/unsave a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleSavedResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, vendor]
        cartName:
          type: string
        category:
          type: string
        city:
          type: string
      required: [email, password, name, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        vendorProfile:
          $ref: "#/components/schemas/VendorSummary"
      required: [token, user]
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        vendorProfile:
          $ref: "#/components/schemas/VendorSummary"
      required: [user]
    UserProfile:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, vendor]
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]
    VendorSummary:
      type: object
      properties:
        id:
          type: number
        cartName:
          type: string
        category:
          type: string
        city:
          type: string
        coverPhoto:
          type: string
        startingPrice:
          type: number
        avgRating:
          type: number
        isActive:
          type: boolean
        subscriptionStatus:
          type: string
          enum: [active, inactive]
        onboardingComplete:
          type: boolean
      required: [id, cartName, category, city, isActive, subscriptionStatus, onboardingComplete]
...
    Booking:
      type: object
      properties:
        id:
          type: number
        customerId:
          type: number
        vendorId:
          type: number
        eventDate:
          type: string
        eventType:
          type: string
        guestCount:
          type: number
        location:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, confirmed, declined]
        vendorNote:
          type: string
        createdAt:
          type: string
        customerName:
          type: string
        vendorCartName:
          type: string
        hasReview:
          type: boolean
      required: [id, customerId, vendorId, eventDate, eventType, guestCount, location, status, createdAt]
    CreateBookingBody:
      type: object
      properties:
        vendorId:
          type: number
        eventDate:
          type: string
        eventType:
          type: string
        guestCount:
          type: number
        location:
          type: string
        message:
          type: string
      required: [vendorId, eventDate, eventType, guestCount, location]
    RespondBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, declined]
        vendorNote:
          type: string
      required: [status]
    BookingResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
      required: [booking]
...
    Review:
      type: object
      properties:
        id:
          type: number
        bookingId:
          type: number
        customerId:
          type: number
        vendorId:
          type: number
        rating:
          type: number
        body:
          type: string
        createdAt:
          type: string
        customerName:
          type: string
      required: [id, bookingId, customerId, vendorId, rating, createdAt]
    CreateReviewBody:
      type: object
      properties:
        bookingId:
          type: number
        rating:
          type: number
        body:
          type: string
      required: [bookingId, rating]
    ReviewResponse:
      type: object
      properties:
        review:
          $ref: "#/components/schemas/Review"
      required: [review]
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
      required: [reviews]
...
    ToggleSavedResponse:
      type: object
      properties:
        saved:
          type: boolean
      required: [saved]
...