# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  customerId: number;
  vendorId: number;
  eventDate: string;
  eventType: string;
  guestCount: number;
  location: string;
  message?: string;
  status: BookingStatus;
  vendorNote?: string;
  createdAt: string;
  customerName?: string;
  vendorCartName?: string;
  hasReview?: boolean;
}