# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  vendorId: number;
  eventDate: string;
  eventType: string;
  guestCount: number;
  location: string;
  message?: string;
}