# lib/api-zod/src/generated/types/respondBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondBookingBodyStatus } from "./respondBookingBodyStatus";

export interface RespondBookingBody {
  status: RespondBookingBodyStatus;
  vendorNote?: string;
}