# lib/api-zod/src/generated/types/vendorSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorSummarySubscriptionStatus } from "./vendorSummarySubscriptionStatus";

export interface VendorSummary {
  id: number;
  cartName: string;
  category: string;
  city: string;
  coverPhoto?: string;
  startingPrice?: number;
  avgRating?: number;
  isActive: boolean;
  subscriptionStatus: VendorSummarySubscriptionStatus;
  onboardingComplete: boolean;
}